seeders

+++ app/Http/Controllers/PermissionController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
 
 class PermissionController extends Controller
@@ -14,6 +15,8 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('permission_index'), 403);
+
         $permissions = Permission::paginate(10);
 
         return view('permissions.index', compact('permissions'));
@@ -26,6 +29,8 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('permission_create'), 403);
+
         return view('permissions.create');
     }
 
@@ -50,6 +55,8 @@ class PermissionController extends Controller
      */
     public function show(Permission $permission)
     {
+        abort_if(Gate::denies('permission_show'), 403);
+
         return view('permissions.show', compact('permission'));
     }
 
@@ -61,6 +68,8 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
+        abort_if(Gate::denies('permission_edit'), 403);
+
         return view('permissions.edit', compact('permission'));
     }
 
@@ -86,6 +95,8 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {
+        abort_if(Gate::denies('permission_delete'), 403);
+
         $permission->delete();
 
         return redirect()->route('permissions.index');
