+++ app/Http/Controllers/PostController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\Post;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
@@ -14,6 +15,8 @@ class PostController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('post_index'), 403);
+
         $posts = Post::paginate(5);
         return view('posts.index', compact('posts'));
     }
@@ -25,6 +28,8 @@ class PostController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('post_create'), 403);
+
         return view('posts.create');
     }
 
@@ -49,6 +54,8 @@ class PostController extends Controller
      */
     public function show(Post $post)
     {
+        abort_if(Gate::denies('post_show'), 403);
+
         return view('posts.show', compact('post'));
     }
 
@@ -60,6 +67,8 @@ class PostController extends Controller
      */
     public function edit(Post $post)
     {
+        abort_if(Gate::denies('post_edit'), 403);
+
         return view('posts.edit', compact('post'));
     }
 
@@ -85,6 +94,8 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
+        abort_if(Gate::denies('post_delete'), 403);
+
         $post->delete();
 
         return redirect()->route('posts.index');
