+++ app/Http/Controllers/RoleController.php
@@ -3,8 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-use Spatie\Permission\Models\Role;
+use Illuminate\Support\Facades\Gate;
 use Spatie\Permission\Models\Permission;
+use Spatie\Permission\Models\Role;
 
 class RoleController extends Controller
 {
@@ -15,6 +16,8 @@ class RoleController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('role_index'), 403);
+
         $roles = Role::paginate(10);
 
         return view('roles.index', compact('roles'));
@@ -27,6 +30,8 @@ class RoleController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('role_create'), 403);
+
         $permissions = Permission::all()->pluck('name', 'id');
         // dd($permissions);
         return view('roles.create', compact('permissions'));
@@ -42,7 +47,8 @@ class RoleController extends Controller
     {
         $role = Role::create($request->only('name'));
 
-        $role->permissions()->syncPermissions($request->input('permissions', []));
+        // $role->permissions()->sync($request->input('permissions', []));
+        $role->syncPermissions($request->input('permissions', []));
 
         return redirect()->route('roles.index');
     }
@@ -55,6 +61,8 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
+        abort_if(Gate::denies('role_show'), 403);
+
         $role->load('permissions');
         return view('roles.show', compact('role'));
     }
@@ -67,6 +75,8 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        abort_if(Gate::denies('role_edit'), 403);
+
         $permissions = Permission::all()->pluck('name', 'id');
         $role->load('permissions');
         // dd($role);
@@ -84,7 +94,8 @@ class RoleController extends Controller
     {
         $role->update($request->only('name'));
 
-        $role->permissions()->syncPermissions($request->input('permissions', []));
+        // $role->permissions()->sync($request->input('permissions', []));
+        $role->syncPermissions($request->input('permissions', []));
 
         return redirect()->route('roles.index');
     }
@@ -97,6 +108,8 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        abort_if(Gate::denies('role_delete'), 403);
+
         $role->delete();
 
         return redirect()->route('roles.index');
